[WIP] Add preloader option. Need to organize the typing
