+++ test.js
@@ -1,0 +1,3 @@
+const { normalize, config } = require('./dist/main-node')
+config.preloadCache = true
+normalize('滋賀県米原市大久保848-1').then(console.log)
